backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
web3.py==6.0.0